Use Float(32-bit) instead of Double(64-bit)

+++ app/src/main/kotlin/com/looker/kenko/data/local/model/SetEntity.kt
@@ -9,7 +9,7 @@ import kotlinx.serialization.Serializable
 @SerialName("set")
 data class SetEntity(
     val repsOrDuration: Int,
-    val weight: Double,
+    val weight: Float,
     val type: SetType,
     val exercise: ExerciseEntity,
 )
